+++ api.py
@@ -6,7 +6,7 @@ from app.user_auth import UserAuth
 from app.user_deauth import UserDeauth
 from app.showtimes import CreateShowtimeHandler, ListShowtimesHandler,\
     ShowtimeAccessTokens
-from app.promotionkeys import PromotionKeysHandler
+from app.promotion_keys import PromotionKeysHandler
 from app.services_auth.google_auth import GoogleAuth
 from app.services_auth.facebook_auth import FacebookAuth
 from app.services_auth.spotify_auth import SpotifyAuth
